requirements.txt
amqp==2.5.0
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
backports.csv==1.0.7
billiard==3.6.0.0
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.3.0
celery-haystack==0.10
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-appconf==1.0.3
django-cors-headers==3.0.2
django-countries==5.3.3
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.19
django-excel==0.0.10
django-grappelli==2.13.1
django-import-export==1.2.0
django-jsonfield==1.2.0
django-nine==0.2.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==3.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djoser==1.7.0
docutils==0.15.2
elastic-apm==5.1.0
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==2.0.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.1
kombu==4.6.3
lazy-object-proxy==1.4.1
lml==0.0.9
lxml==4.4.0
MarkupSafe==1.1.1
mysqlclient==1.4.2
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
parso==0.5.1
pdfrw==0.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.19
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
python-dotenv==0.10.3
python-magic==0.4.15
python3-logstash==0.4.80
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
text-unidecode==1.2
texttable==1.6.2
tinycss2==1.0.2
traitlets==4.3.2
typed-ast==1.4.0
typing==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==48
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
